udpate

+++ engineering/Project.Maintainer/Maintenance.cs
@@ -9,14 +9,31 @@ public abstract class Maintenance
 {
     public abstract string Paths();
 
+    public const string StartColor = "\e[32;45m";
+    public const string StopColor = "\e[0m";
+
     public void Maintain()
     {
         var paths = Paths();
         var src = Path.Combine(Program.UpdateContentDirectory, paths);
         var dst = Path.Combine(Program.RepoDirectory, paths);
         Helper.ShallowCheckout(Program.UpdateContentDirectory, paths);
+
+        if (File.Exists(src))
+        {
+            File.Delete(dst);
+            File.Copy(src,dst);
+            Console.WriteLine($"{StartColor}copy file from {src} to {dst}{StopColor}");
+        }
+
+        if (!Directory.Exists(src))
+        {
+            return;
+        }
+
         Directory.Delete(dst);
         Directory.Move(src,
                        dst);
+        Console.WriteLine($"{StartColor}copy directory from {src} to {dst}{StopColor}");
     }
 }
